now  we  can talking about  a real project ! please download and test <<by said>>

+++ plateform/src/org/java/mql/dao/EtudiantDao.java
@@ -6,9 +6,9 @@ import org.java.mql.dao.mediator.DAOservice;
 import org.java.mql.models.Etudiant;
 
 public interface EtudiantDao extends DAOservice{
-	long addEtudiant(Etudiant etudiant);
+	int addEtudiant(Etudiant etudiant);
 	Etudiant deleteEtudiant(Etudiant etudiant);
 	List<Etudiant> selectAllEtudiant();
 	Etudiant selectEtudiantById(long id);
-	long updateEtudiant(Etudiant etudiant);
+	int updateEtudiant(Etudiant etudiant);
 }
